pyproject.toml
# Project configuration for Drone Fleet Search Control System
[project]
name = "drone-fleet-search"
version = "0.1.0"
description = "AI-driven drone fleet search control system using AWS and ArduPilot"
authors = [{name = "Chris Creel"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "boto3>=1.35",
    "aws-cdk-lib>=2.0",
    "constructs>=10.0",
]

[project.optional-dependencies]
cloud = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "boto3>=1.35",
]
edge = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pymavlink>=2.4",
    "paho-mqtt>=2.0",
    "numpy>=1.26",
    "Pillow>=10.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "pyright>=1.1",
    "vulture>=2.11",
    "bandit>=1.7",
    "pip-audit>=2.0",
    "pre-commit>=3.0",
    "commitizen>=3.13",
    "moto[dynamodb,s3,iot,sqs,iam,lambda,apigateway,events,logs,sns,cognito]>=5.0",
    "boto3-stubs[dynamodb,s3,iot-data,sqs,iam,lambda,apigateway,events,logs,sns,cognito,bedrock-runtime]>=1.35",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
include = ["src", "infra"]
exclude = ["**/__pycache__"]
reportMissingImports = "warning"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "infra", "edge"]

[tool.ruff.lint]
select = [
    # Core
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade

    # Code quality
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "D",      # pydocstyle
    "ERA",    # eradicate (commented-out code)
    "BLE",    # flake8-blind-except

    # Security
    "S",      # flake8-bandit (security checks in linting)

    # Modern Python
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
    "PERF",   # Performance anti-patterns
    "PTH",    # pathlib preference over os.path

    # Error handling (catches defensive patterns)
    "TRY",    # tryceratops (exception handling)
    "EM",     # error messages in exceptions

    # Function design
    "ARG",    # unused arguments
    "FBT",    # boolean-trap (boolean positional args)
    "PLR",    # pylint refactor (includes complexity)
    "PLC",    # pylint convention
    "PLE",    # pylint error
    "PLW",    # pylint warning

    # Type checking related
    "TCH",    # type-checking imports
    "ANN",    # type annotations

    # Code hygiene
    "T20",    # flake8-print (no print statements)
    "PIE",    # misc lints
    "RSE",    # raise statements
    "RET",    # return statements
    "SLF",    # private member access
    "ICN",    # import conventions
    "PGH",    # pygrep-hooks
]

ignore = [
    # Docstrings - allow flexibility
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "D107",   # Missing docstring in __init__

    # Type annotations - some flexibility
    "ANN401", # Dynamically typed expressions (Any) - handled by pyright

    # PEP 695 type parameters
    "UP046",  # Generic class uses Generic subclass instead of type parameters

    # Allow assert in tests (S101 is flake8-bandit assert)
    "S101",   # Use of assert - OK in tests, caught by per-file-ignores

    # Function complexity - handled by CodeRabbit with more context
    "PLR0913", # Too many arguments (>5)

    # Sometimes needed for API compatibility
    "ARG002", # Unused method argument (sometimes needed for interface compliance)

    # Error messages - sometimes not practical
    "EM101",  # Exception must not use string literal
    "EM102",  # Exception must not use f-string literal

    # Try-except - some patterns are valid
    "TRY003", # Avoid specifying long messages outside exception class
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # Assert allowed in tests
    "ARG",     # Unused args OK (fixtures)
    "PLR2004", # Magic values OK in tests
    "ANN",     # Type annotations optional in tests
    "D",       # Docstrings optional in tests
    "TCH",     # Type-checking imports OK at runtime in tests
]
"edge_tests/**/*.py" = [
    "S101",    # Assert allowed in tests
    "ARG",     # Unused args OK (fixtures)
    "PLR2004", # Magic values OK in tests
    "ANN",     # Type annotations optional in tests
    "D",       # Docstrings optional in tests
    "TCH",     # Type-checking imports OK at runtime in tests
]
"infra_tests/**/*.py" = [
    "S101",    # Assert allowed in tests
    "ARG",     # Unused args OK (fixtures)
    "PLR2004", # Magic values OK in tests
    "ANN",     # Type annotations optional in tests
    "D",       # Docstrings optional in tests
    "TCH",     # Type-checking imports OK at runtime in tests
]
"infra/**/*.py" = [
    "ARG001",  # Unused CDK construct args (required by CDK)
    "D",       # Docstrings optional in infra
]
"edge/**/*.py" = [
    "D",       # Docstrings optional in edge (separate from cloud docs)
]
"scripts/**/*.py" = [
    "T20",     # Print allowed in scripts
    "S",       # Security checks less strict in scripts
    "ANN",     # Type annotations less strict in scripts
    "D",       # Docstrings optional in scripts
    "PTH",     # pathlib not required in scripts
    "TRY",     # try-except patterns relaxed in scripts
    "RUF012",  # Mutable class attributes OK in scripts
    "PGH003",  # Specific rule codes not required in scripts
    "PIE810",  # startswith/endswith tuple not required
    "E501",    # Line length relaxed in scripts
]
"integration_tests/**/*.py" = [
    "S101",    # Assert allowed in tests
    "ARG",     # Unused args OK
    "PLR2004", # Magic values OK in tests
    "ANN",     # Type annotations optional in tests
    "D",       # Docstrings optional in tests
    "TCH",     # Type-checking imports OK at runtime in tests
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["src", "edge"]

[tool.ruff.lint.pylint]
max-args = 5
max-branches = 12
max-returns = 6
max-statements = 50

[tool.ruff.lint.flake8-type-checking]
strict = true

[tool.vulture]
exclude = ["src/models/", "edge/models/"]
min_confidence = 80

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-fail-under=95 --cov-report=term-missing"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
branch = true
source = ["src"]

# Commitizen - Conventional Commits
[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
version_files = [
    "pyproject.toml:^version",
]
update_changelog_on_bump = true

[tool.commitizen.customize]
message_template = "{{change_type}}{% if scope %}({{scope}}){% endif %}: {{subject}}"
schema_pattern = "^(feat|fix|docs|style|refactor|perf|test|build|ci|chore|revert)(\\([a-z0-9-]+\\))?!?:\\s.+"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
